AMAP in cents (#496)
Keeps amounts, accounts and balances in cents and calculates them in
cents, i.e. in int. Then the front-end displays in euros and takes in
euros.

---------

+++ lib/amap/providers/cash_list_provider.dart
@@ -18,7 +18,7 @@ class CashListProvider extends ListNotifier<Cash> {
     return await add(cashRepository.createCash, cash);
   }
 
-  Future<bool> updateCash(Cash addedCash, double previousCashAmount) async {
+  Future<bool> updateCash(Cash addedCash, int previousCashAmount) async {
     return await update(
       cashRepository.updateCash,
       (cashList, c) => cashList
